Refactoring mesh reading capability into an I/O layer
