+++ framework/mesh/mesh_generator/extruder_mesh_generator.h
@@ -23,8 +23,8 @@ public:
   explicit ExtruderMeshGenerator(const InputParameters& params);
 
 protected:
-  std::unique_ptr<UnpartitionedMesh>
-  GenerateUnpartitionedMesh(std::unique_ptr<UnpartitionedMesh> input_umesh) override;
+  std::shared_ptr<UnpartitionedMesh>
+  GenerateUnpartitionedMesh(std::shared_ptr<UnpartitionedMesh> input_umesh) override;
 
   const std::string top_boundary_name_;
   const std::string bottom_boundary_name_;
