+++ framework/mesh/mesh_generator/from_file_mesh_generator.h
@@ -15,8 +15,8 @@ public:
   explicit FromFileMeshGenerator(const InputParameters& params);
 
 protected:
-  std::unique_ptr<UnpartitionedMesh>
-  GenerateUnpartitionedMesh(std::unique_ptr<UnpartitionedMesh> input_umesh) override;
+  std::shared_ptr<UnpartitionedMesh>
+  GenerateUnpartitionedMesh(std::shared_ptr<UnpartitionedMesh> input_umesh) override;
   const std::string filename_;
   const std::string material_id_fieldname_;
   const std::string boundary_id_fieldname_;
