+++ framework/mesh/mesh_generator/mesh_generator.cc
@@ -70,8 +70,8 @@ MeshGenerator::MeshGenerator(const InputParameters& params)
   partitioner_ = &GetStackItem<GraphPartitioner>(object_stack, partitioner_handle, __FUNCTION__);
 }
 
-std::unique_ptr<UnpartitionedMesh>
-MeshGenerator::GenerateUnpartitionedMesh(std::unique_ptr<UnpartitionedMesh> input_umesh)
+std::shared_ptr<UnpartitionedMesh>
+MeshGenerator::GenerateUnpartitionedMesh(std::shared_ptr<UnpartitionedMesh> input_umesh)
 {
   return input_umesh;
 }
@@ -81,15 +81,15 @@ MeshGenerator::Execute()
 {
   // Execute all input generators
   // Note these could be empty
-  std::unique_ptr<UnpartitionedMesh> current_umesh = nullptr;
+  std::shared_ptr<UnpartitionedMesh> current_umesh = nullptr;
   for (auto mesh_generator_ptr : inputs_)
   {
-    auto new_umesh = mesh_generator_ptr->GenerateUnpartitionedMesh(std::move(current_umesh));
-    current_umesh = std::move(new_umesh);
+    auto new_umesh = mesh_generator_ptr->GenerateUnpartitionedMesh(current_umesh);
+    current_umesh = new_umesh;
   }
 
   // Generate final umesh and convert it
-  current_umesh = GenerateUnpartitionedMesh(std::move(current_umesh));
+  current_umesh = GenerateUnpartitionedMesh(current_umesh);
 
   std::vector<int64_t> cell_pids;
   if (opensn::mpi_comm.rank() == 0)
@@ -151,7 +151,7 @@ MeshGenerator::ComputeAndPrintStats(const MeshContinuum& grid)
 std::vector<int64_t>
 MeshGenerator::PartitionMesh(const UnpartitionedMesh& input_umesh, int num_partitions)
 {
-  const auto& raw_cells = input_umesh.GetRawCells();
+  const auto& raw_cells = input_umesh.RawCells();
   const size_t num_raw_cells = raw_cells.size();
 
   OpenSnLogicalErrorIf(num_raw_cells == 0, "No cells in final input mesh");
@@ -206,17 +206,17 @@ MeshGenerator::PartitionMesh(const UnpartitionedMesh& input_umesh, int num_parti
 }
 
 std::shared_ptr<MeshContinuum>
-MeshGenerator::SetupMesh(std::unique_ptr<UnpartitionedMesh> input_umesh_ptr,
+MeshGenerator::SetupMesh(std::shared_ptr<UnpartitionedMesh> input_umesh,
                          const std::vector<int64_t>& cell_pids)
 {
   // Convert mesh
   auto grid_ptr = MeshContinuum::New();
 
-  grid_ptr->GetBoundaryIDMap() = input_umesh_ptr->GetMeshOptions().boundary_id_map;
+  grid_ptr->GetBoundaryIDMap() = input_umesh->BoundaryIDMap();
 
-  auto& vertex_subs = input_umesh_ptr->GetVertextCellSubscriptions();
+  auto& vertex_subs = input_umesh->GetVertextCellSubscriptions();
   size_t cell_globl_id = 0;
-  for (auto& raw_cell : input_umesh_ptr->GetRawCells())
+  for (auto& raw_cell : input_umesh->RawCells())
   {
     if (CellHasLocalScope(
           opensn::mpi_comm.rank(), *raw_cell, cell_globl_id, vertex_subs, cell_pids))
@@ -224,10 +224,10 @@ MeshGenerator::SetupMesh(std::unique_ptr<UnpartitionedMesh> input_umesh_ptr,
       auto cell = SetupCell(*raw_cell,
                             cell_globl_id,
                             cell_pids[cell_globl_id],
-                            STLVertexListHelper(input_umesh_ptr->GetVertices()));
+                            STLVertexListHelper(input_umesh->Vertices()));
 
       for (uint64_t vid : cell->vertex_ids_)
-        grid_ptr->vertices.Insert(vid, input_umesh_ptr->GetVertices()[vid]);
+        grid_ptr->vertices.Insert(vid, input_umesh->Vertices()[vid]);
 
       grid_ptr->cells.push_back(std::move(cell));
     }
@@ -239,12 +239,12 @@ MeshGenerator::SetupMesh(std::unique_ptr<UnpartitionedMesh> input_umesh_ptr,
   } // for raw_cell
 
   SetGridAttributes(*grid_ptr,
-                    input_umesh_ptr->GetMeshAttributes(),
-                    {input_umesh_ptr->GetMeshOptions().ortho_Nx,
-                     input_umesh_ptr->GetMeshOptions().ortho_Ny,
-                     input_umesh_ptr->GetMeshOptions().ortho_Nz});
+                    input_umesh->Attributes(),
+                    {input_umesh->MeshOptions().ortho_Nx,
+                     input_umesh->MeshOptions().ortho_Ny,
+                     input_umesh->MeshOptions().ortho_Nz});
 
-  grid_ptr->SetGlobalVertexCount(input_umesh_ptr->GetVertices().size());
+  grid_ptr->SetGlobalVertexCount(input_umesh->Vertices().size());
 
   ComputeAndPrintStats(*grid_ptr);
