+++ framework/mesh/mesh_generator/mesh_generator.h
@@ -45,13 +45,11 @@ public:
 
   /**
    * Virtual method to generate the unpartitioned mesh for the next step.
-   */
-  virtual std::unique_ptr<UnpartitionedMesh>
-
-  /**
+   *
    * Default behavior here is to return the input umesh unaltered.
    */
-  GenerateUnpartitionedMesh(std::unique_ptr<UnpartitionedMesh> input_umesh);
+  virtual std::shared_ptr<UnpartitionedMesh>
+  GenerateUnpartitionedMesh(std::shared_ptr<UnpartitionedMesh> input_umesh);
 
   struct VertexListHelper
   {
@@ -75,7 +73,7 @@ protected:
   /**
    * Executes the partitioner and configures the mesh as a real mesh.
    */
-  std::shared_ptr<MeshContinuum> SetupMesh(std::unique_ptr<UnpartitionedMesh> input_umesh_ptr,
+  std::shared_ptr<MeshContinuum> SetupMesh(std::shared_ptr<UnpartitionedMesh> input_umesh,
                                            const std::vector<int64_t>& cell_pids);
 
   /**
