+++ framework/mesh/mesh_generator/orthogonal_mesh_generator.cc
@@ -92,11 +92,11 @@ OrthogonalMeshGenerator::OrthogonalMeshGenerator(const InputParameters& params)
                             "values to be monotonically increasing. Node set: " +
                             outstr.str());
     }
-  } // for node_set in node_sets_
+  }
 }
 
-std::unique_ptr<UnpartitionedMesh>
-OrthogonalMeshGenerator::GenerateUnpartitionedMesh(std::unique_ptr<UnpartitionedMesh> input_umesh)
+std::shared_ptr<UnpartitionedMesh>
+OrthogonalMeshGenerator::GenerateUnpartitionedMesh(std::shared_ptr<UnpartitionedMesh> input_umesh)
 {
   OpenSnInvalidArgumentIf(input_umesh != nullptr,
                           "OrthogonalMeshGenerator can not be preceded by another"
@@ -109,14 +109,14 @@ OrthogonalMeshGenerator::GenerateUnpartitionedMesh(std::unique_ptr<Unpartitioned
   else if (node_sets_.size() == 3)
     return CreateUnpartitioned3DOrthoMesh(node_sets_[0], node_sets_[1], node_sets_[2]);
   else
-    throw std::logic_error(
-      ""); // This will never get triggered because of the checks in constructor
+    // This will never get triggered because of the checks in constructor
+    throw std::logic_error("");
 }
 
-std::unique_ptr<UnpartitionedMesh>
+std::shared_ptr<UnpartitionedMesh>
 OrthogonalMeshGenerator::CreateUnpartitioned1DOrthoMesh(const std::vector<double>& vertices)
 {
-  auto umesh = std::make_unique<UnpartitionedMesh>();
+  auto umesh = std::make_shared<UnpartitionedMesh>();
 
   // Reorient 1D verts along z
   std::vector<Vertex> zverts;
@@ -124,20 +124,20 @@ OrthogonalMeshGenerator::CreateUnpartitioned1DOrthoMesh(const std::vector<double
   for (double z_coord : vertices)
     zverts.emplace_back(0.0, 0.0, z_coord);
 
-  umesh->GetMeshAttributes() = DIMENSION_1 | ORTHOGONAL;
+  umesh->Attributes() = DIMENSION_1 | ORTHOGONAL;
 
   // Create vertices
   size_t Nz = vertices.size();
 
-  umesh->GetMeshOptions().ortho_Nx = 1;
-  umesh->GetMeshOptions().ortho_Ny = 1;
-  umesh->GetMeshOptions().ortho_Nz = Nz - 1;
-  umesh->GetMeshOptions().boundary_id_map[4] = "ZMAX";
-  umesh->GetMeshOptions().boundary_id_map[5] = "ZMIN";
+  umesh->MeshOptions().ortho_Nx = 1;
+  umesh->MeshOptions().ortho_Ny = 1;
+  umesh->MeshOptions().ortho_Nz = Nz - 1;
+  umesh->AddBoundary(4, "ZMAX");
+  umesh->AddBoundary(5, "ZMIN");
 
-  umesh->GetVertices().reserve(Nz);
+  umesh->Vertices().reserve(Nz);
   for (auto& vertex : zverts)
-    umesh->GetVertices().push_back(vertex);
+    umesh->Vertices().push_back(vertex);
 
   // Create cells
   const size_t max_cz = zverts.size() - 2;
@@ -185,29 +185,29 @@ OrthogonalMeshGenerator::CreateUnpartitioned1DOrthoMesh(const std::vector<double
   return umesh;
 }
 
-std::unique_ptr<UnpartitionedMesh>
+std::shared_ptr<UnpartitionedMesh>
 OrthogonalMeshGenerator::CreateUnpartitioned2DOrthoMesh(const std::vector<double>& vertices_1d_x,
                                                         const std::vector<double>& vertices_1d_y)
 {
-  auto umesh = std::make_unique<UnpartitionedMesh>();
+  auto umesh = std::make_shared<UnpartitionedMesh>();
 
-  umesh->GetMeshAttributes() = DIMENSION_2 | ORTHOGONAL;
+  umesh->Attributes() = DIMENSION_2 | ORTHOGONAL;
 
   // Create vertices
   const size_t Nx = vertices_1d_x.size();
   const size_t Ny = vertices_1d_y.size();
 
-  umesh->GetMeshOptions().ortho_Nx = Nx - 1;
-  umesh->GetMeshOptions().ortho_Ny = Ny - 1;
-  umesh->GetMeshOptions().ortho_Nz = 1;
-  umesh->GetMeshOptions().boundary_id_map[0] = "XMAX";
-  umesh->GetMeshOptions().boundary_id_map[1] = "XMIN";
-  umesh->GetMeshOptions().boundary_id_map[2] = "YMAX";
-  umesh->GetMeshOptions().boundary_id_map[3] = "YMIN";
+  umesh->MeshOptions().ortho_Nx = Nx - 1;
+  umesh->MeshOptions().ortho_Ny = Ny - 1;
+  umesh->MeshOptions().ortho_Nz = 1;
+  umesh->AddBoundary(0, "XMAX");
+  umesh->AddBoundary(1, "XMIN");
+  umesh->AddBoundary(2, "YMAX");
+  umesh->AddBoundary(3, "YMIN");
 
   typedef std::vector<uint64_t> VecIDs;
   std::vector<VecIDs> vertex_ij_to_i_map(Ny, VecIDs(Nx));
-  umesh->GetVertices().reserve(Nx * Ny);
+  umesh->Vertices().reserve(Nx * Ny);
   {
     uint64_t k = 0;
     for (size_t i = 0; i < Ny; ++i)
@@ -215,9 +215,9 @@ OrthogonalMeshGenerator::CreateUnpartitioned2DOrthoMesh(const std::vector<double
       for (size_t j = 0; j < Nx; ++j)
       {
         vertex_ij_to_i_map[i][j] = k++;
-        umesh->GetVertices().emplace_back(vertices_1d_x[j], vertices_1d_y[i], 0.0);
-      } // for j
-    }   // for i
+        umesh->Vertices().emplace_back(vertices_1d_x[j], vertices_1d_y[i], 0.0);
+      }
+    }
   }
 
   std::vector<VecIDs> cells_ij_to_i_map(Ny - 1, VecIDs(Nx - 1));
@@ -298,8 +298,8 @@ OrthogonalMeshGenerator::CreateUnpartitioned2DOrthoMesh(const std::vector<double
       }
 
       umesh->AddCell(cell);
-    } // for j
-  }   // for i
+    }
+  }
 
   umesh->ComputeCentroidsAndCheckQuality();
   umesh->BuildMeshConnectivity();
@@ -307,29 +307,29 @@ OrthogonalMeshGenerator::CreateUnpartitioned2DOrthoMesh(const std::vector<double
   return umesh;
 }
 
-std::unique_ptr<UnpartitionedMesh>
+std::shared_ptr<UnpartitionedMesh>
 OrthogonalMeshGenerator::CreateUnpartitioned3DOrthoMesh(const std::vector<double>& vertices_1d_x,
                                                         const std::vector<double>& vertices_1d_y,
                                                         const std::vector<double>& vertices_1d_z)
 {
-  auto umesh = std::make_unique<UnpartitionedMesh>();
+  auto umesh = std::make_shared<UnpartitionedMesh>();
 
-  umesh->GetMeshAttributes() = DIMENSION_3 | ORTHOGONAL;
+  umesh->Attributes() = DIMENSION_3 | ORTHOGONAL;
 
   // Create vertices
   size_t Nx = vertices_1d_x.size();
   size_t Ny = vertices_1d_y.size();
   size_t Nz = vertices_1d_z.size();
 
-  umesh->GetMeshOptions().ortho_Nx = Nx - 1;
-  umesh->GetMeshOptions().ortho_Ny = Ny - 1;
-  umesh->GetMeshOptions().ortho_Nz = Nz - 1;
-  umesh->GetMeshOptions().boundary_id_map[0] = "XMAX";
-  umesh->GetMeshOptions().boundary_id_map[1] = "XMIN";
-  umesh->GetMeshOptions().boundary_id_map[2] = "YMAX";
-  umesh->GetMeshOptions().boundary_id_map[3] = "YMIN";
-  umesh->GetMeshOptions().boundary_id_map[4] = "ZMAX";
-  umesh->GetMeshOptions().boundary_id_map[5] = "ZMIN";
+  umesh->MeshOptions().ortho_Nx = Nx - 1;
+  umesh->MeshOptions().ortho_Ny = Ny - 1;
+  umesh->MeshOptions().ortho_Nz = Nz - 1;
+  umesh->AddBoundary(0, "XMAX");
+  umesh->AddBoundary(1, "XMIN");
+  umesh->AddBoundary(2, "YMAX");
+  umesh->AddBoundary(3, "YMIN");
+  umesh->AddBoundary(4, "ZMAX");
+  umesh->AddBoundary(5, "ZMIN");
 
   // i is j, and j is i, MADNESS explanation:
   // In math convention the i-index refers to the ith row
@@ -342,7 +342,7 @@ OrthogonalMeshGenerator::CreateUnpartitioned3DOrthoMesh(const std::vector<double
   for (auto& vec : vertex_ijk_to_i_map)
     vec.resize(Nx, VecIDs(Nz));
 
-  umesh->GetVertices().reserve(Nx * Ny * Nz);
+  umesh->Vertices().reserve(Nx * Ny * Nz);
   {
     uint64_t c = 0;
     for (size_t i = 0; i < Ny; ++i)
@@ -352,10 +352,10 @@ OrthogonalMeshGenerator::CreateUnpartitioned3DOrthoMesh(const std::vector<double
         for (size_t k = 0; k < Nz; ++k)
         {
           vertex_ijk_to_i_map[i][j][k] = c++;
-          umesh->GetVertices().emplace_back(vertices_1d_x[j], vertices_1d_y[i], vertices_1d_z[k]);
-        } // for k
-      }   // for j
-    }     // for i
+          umesh->Vertices().emplace_back(vertices_1d_x[j], vertices_1d_y[i], vertices_1d_z[k]);
+        }
+      }
+    }
   }
 
   std::vector<VecVecIDs> cells_ijk_to_i_map(Ny - 1);
@@ -463,9 +463,9 @@ OrthogonalMeshGenerator::CreateUnpartitioned3DOrthoMesh(const std::vector<double
         }
 
         umesh->AddCell(cell);
-      } // for k
-    }   // for j
-  }     // for i
+      }
+    }
+  }
 
   umesh->ComputeCentroidsAndCheckQuality();
   umesh->BuildMeshConnectivity();
